+++ src/Repositories/ThreadRepository.php
@@ -5,7 +5,6 @@ namespace RTippin\Messenger\Repositories;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Collection;
 use RTippin\Messenger\Messenger;
-use RTippin\Messenger\Models\Participant;
 use RTippin\Messenger\Models\Thread;
 use RTippin\Messenger\Support\Helpers;
 
@@ -31,7 +30,7 @@ class ThreadRepository
      */
     public function getProviderThreadsIndex(): Collection
     {
-        return Thread::hasProvider('participants', $this->messenger->getProvider())
+        return Thread::hasProvider($this->messenger->getProvider())
             ->latest('updated_at')
             ->with([
                 'participants.owner',
@@ -48,7 +47,7 @@ class ThreadRepository
      */
     public function getProviderThreadsPage(Thread $thread): Collection
     {
-        return Thread::hasProvider('participants', $this->messenger->getProvider())
+        return Thread::hasProvider($this->messenger->getProvider())
             ->latest('updated_at')
             ->with([
                 'participants.owner',
@@ -66,7 +65,7 @@ class ThreadRepository
      */
     public function getProviderThreadsWithActiveCallsBuilder(): Builder
     {
-        return Thread::hasProvider('participants', $this->messenger->getProvider())->has('activeCall');
+        return Thread::hasProvider($this->messenger->getProvider())->has('activeCall');
     }
 
     /**
@@ -82,7 +81,7 @@ class ThreadRepository
      */
     public function getProviderOldestThread(): ?Thread
     {
-        return Thread::hasProvider('participants', $this->messenger->getProvider())
+        return Thread::hasProvider($this->messenger->getProvider())
             ->oldest('updated_at')
             ->first();
     }
@@ -92,13 +91,10 @@ class ThreadRepository
      */
     public function getProviderUnreadThreadsBuilder(): Builder
     {
-        return Thread::whereHas('participants', function (Builder $query) {
-            /** @var Builder|Participant $query */
-            return $query->forProvider($this->messenger->getProvider())
-                ->where(function (Builder $query) {
-                    return $query->whereNull('participants.last_read')
-                        ->orWhere('threads.updated_at', '>', $query->raw('participants.last_read'));
-                });
-        });
+        return Thread::hasProvider($this->messenger->getProvider())
+            ->where(function (Builder $query) {
+                return $query->whereNull('participants.last_read')
+                    ->orWhere('threads.updated_at', '>', $query->raw('participants.last_read'));
+            });
     }
 }