+++ tests/Messenger/ThreadLocatorServiceTest.php
@@ -27,6 +27,29 @@ class ThreadLocatorServiceTest extends FeatureTestCase
         $results = $this->locator->setAlias('user')->setId($this->doe->getKey())->locate();
 
         $this->assertSame($this->doe->getKey(), $results->getRecipient()->getKey());
+        $this->assertSame($this->doe->getMorphClass(), $results->getRecipient()->getMorphClass());
+        $this->assertSame($thread->id, $results->getThread()->id);
+    }
+
+    /** @test */
+    public function it_returns_user_without_existing_thread_id()
+    {
+        $results = $this->locator->setAlias('user')->setId($this->doe->getKey())->locate();
+
+        $this->assertSame($this->doe->getKey(), $results->getRecipient()->getKey());
+        $this->assertSame($this->doe->getMorphClass(), $results->getRecipient()->getMorphClass());
+        $this->assertNull($results->getThread());
+    }
+
+    /** @test */
+    public function it_returns_company_with_existing_thread_id()
+    {
+        $thread = $this->createPrivateThread($this->tippin, $this->developers);
+
+        $results = $this->locator->setAlias('company')->setId($this->developers->getKey())->locate();
+
+        $this->assertSame($this->developers->getKey(), $results->getRecipient()->getKey());
+        $this->assertSame($this->developers->getMorphClass(), $results->getRecipient()->getMorphClass());
         $this->assertSame($thread->id, $results->getThread()->id);
     }
 
@@ -36,6 +59,7 @@ class ThreadLocatorServiceTest extends FeatureTestCase
         $results = $this->locator->setAlias('company')->setId($this->developers->getKey())->locate();
 
         $this->assertSame($this->developers->getKey(), $results->getRecipient()->getKey());
+        $this->assertSame($this->developers->getMorphClass(), $results->getRecipient()->getMorphClass());
         $this->assertNull($results->getThread());
     }